Further Optimize+Simplify Snapshot Delete
This commit brings delete code in alignment with
the approach taken in #46250.

There is only one functional change here: index
metadata is now deleted in the same bulk request that
deletes all other unreferenced blobs.
This cleanly separates the first part of the snapshot delete,
the deleting from the root-level metadata from the
other steps that may not propagate an exception to the
`listener`.