Fix GCS Mock Range Downloads
We were not correctly respecting the download range which lead
to the GCS SDK client closing the connection at times.
Also, we are randomly chosing to fail requests in these tests
but set the retry count to `0` by force which could lead to
issues when a chunk of a ranged download is failed and the
retrying input stream doesn't retry the chunk (GCS SDK strangely does not retry
this case on a `5xx`).

Closes #51446