Make CompressedXContent (De-)serialization Cheaper
When working ith `CompressedXContent` we often have to detect the
concrete content type when deserializing durign performance critical logic
like working with mappings or templates on the CS threads.

By adding the source type to the network messages and determining it up front when
we already have the uncompressed bytes available anyway we rarely have to actually run
the content type determination which can be a non-trivial part of deserializing smaller
mappings or templates.