Make ActionListener#wrap more Memory Efficient
The pattern of using `ActionListener.wrap` with another listener as exception
handler by passing the exception consumer `listener::onFailure` is widespread.
This causes some needless indirection and more objects than necessary to be allocated
due to the extra `Consumer<Exception>` this creates.
More importantly though, in situations where we pass listeners created through `.wrap`
calls down the execution and wrap them multiple times, this causes us to retain the
`onResponse` consumers with a lot of wrapping in many cases where we only need the initial
failure handling listener. The consumers can (as in the case of getting snapshot repository
metadata capture objects of non-trivial size that should be GCed asap).

=> this PR adds a way to call `.wrap` on listener instances to build functionally equivalent listeners
that chain in more predictable ways memory wise and are easier to GC.