Add Indirection for Collection #copyOf Methods to Enable Transparent Backports
We have our own homebrew versions of these methods for JDK8 but the fact that they reside on
classes of the same name but different package as their JDK counterparts makes it impossible
to use them for transparent backporting in most cases because the JDK version is already imported.
This adds the neccessary indirection to fix this.