Speed up JsonXContentParser token conversion and text()
This reduces the compile size of convert token by about 20 bytes making it inline in a couple
of places that wouldn't inline before once `unknownTokenException` is deoptimized as dead code.

Same for the change to `text()` that saves some method size there so it inlines into the much used
textOrNull here and there.