Make Mapper building cheaper
This code is still hitting us a lot when running the many-shards benchmarks.
Iterating over the parameters is one of the big slow points. By moving to an
array we get faster iteration and save the multimorphic call site from
creating different kinds of lists before iterating which helps a little.
Also makes the multi fields a little cheaper.